Separate tables for account, PIC, and bank details
This separation allows us to set required fields as NOT NULL
while still allowing each component to be uploaded separately.

The old onboarding table names now refer to views that join the
corresponding underlying tables. This makes it very easy to tell
if a user has been completely onboarded, since they must have
entries in all of the corresponding tables to show up in the view.